requirements.txt
aiohttp
Red-DiscordBot